build auth method in base client instead
